save method created

+++ Backend/ecommerce/src/main/java/com/brq/ecommerce/repositories/ItemPedidoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.brq.ecommerce.models.ItemPedidoModel;
 
 @Repository
-public interface ItemPedidoRepository extends JpaRepository<ItemPedidoModel, Long> {
+public interface ItemPedidoRepository extends JpaRepository<ItemPedidoModel, Integer> {
 
 	
 }
